+++ pkg/cmd/doc.go
@@ -1,0 +1,2 @@
+// Package cmd contains helper functionality to run various okctl commands.
+package cmd
